crud sucursal

+++ src/main/java/com/paraisocanino/hospital_veterinario/repository/SucursalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SucursalRepository extends JpaRepository<Sucursal, Long> {
+public interface SucursalRepository extends JpaRepository<Sucursal, Integer> {
 }
